% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data_GUI.R
\name{input_data_GUI}
\alias{input_data_GUI}
\title{Input Data GUI}
\usage{
input_data_GUI()
}
\value{
Data input user-interface for \emph{cellcount} image analysis
}
\description{
This function provides a Shiny user-interface to easily input data needed for fluorescent microscopy image analysis
outlined in \emph{cellcount}. This Shiny app contains several tabs for data input, directory selection, etc. Users must close
the app by selecting the 'Completion' tab followed by 'Submit data entries' for all entered data to be exported into
the local R environment, where users can then utilize the 'analyze_images_GUI' function to run the \emph{cellcount} image analysis.
}
