% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapped.R
\name{mapped}
\alias{mapped}
\title{Matrix mapping conversion}
\usage{
mapped(x, threshold = 0.3)
}
\arguments{
\item{x}{a grey scale image variable}

\item{threshold}{a numerical value of pixel intensity}
}
\value{
Converts an grey scale image to a data matrix, in a pixel by

pixel format
}
\description{
Takes a grey-scale/8-bit image to convert to a data matrix,
where data under the numerical pixel intensity threshold is removed.
This function is needed for the use of the "image_convert" function.
}
\examples{
images <- list.files("C:/Users/Desktop/Test_Images/",pattern = "tif", full.name = T)
grey_images <- greyscale(images,contrast=2)
mapped(grey_images,threshold=0.3)
}
